+++ src/ska_ser_namespace_manager/core/notifier.py
@@ -85,6 +85,7 @@ class Notifier:
         if status in [
             NamespaceStatus.FAILED.value,
             NamespaceStatus.STALE.value,
+            NamespaceStatus.CANCELLED.value,
         ]:
             return random.choice(
                 [